+++ src/sphinx_autodoc_typehints/_parser.py
@@ -2,7 +2,8 @@
 
 from __future__ import annotations
 
-from typing import TYPE_CHECKING
+import re
+from typing import TYPE_CHECKING, Final
 
 from docutils.utils import new_document
 from sphinx.parsers import RSTParser
@@ -13,6 +14,14 @@ if TYPE_CHECKING:
     from docutils import nodes
     from docutils.frontend import Values
     from docutils.statemachine import StringList
+
+_UNESCAPE_RE: Final = re.compile(
+    r"""
+    \\          # literal backslash
+    ([^ ])      # followed by any non-space character (captured)
+    """,
+    re.VERBOSE,
+)
 
 
 class _RstSnippetParser(RSTParser):
@@ -28,3 +37,11 @@ def parse(inputstr: str, settings: Values | optparse.Values) -> nodes.document:
     # losing the external+ roles the read phase resolves (#753)
     _RstSnippetParser().parse(inputstr, doc)
     return doc
+
+
+def unescape(escaped: str) -> str:
+    escaped = escaped.replace("\x00", "")
+    return _UNESCAPE_RE.sub(r"\1", escaped)
+
+
+__all__ = ["parse", "unescape"]
